test(e2e): split Discord workflow boundary
